// migrations/snapshots/da34d70bb5a0db7a7b160d7a75db8e7997473c38493c72facacca6698ed3a010/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Guest: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly fullName: CodecTypes['pg/text@1']['output'];
      readonly phone: CodecTypes['pg/text@1']['output'] | null;
      readonly side: CodecTypes['pg/text@1']['output'];
      readonly rsvpStatus: CodecTypes['pg/text@1']['output'];
      readonly ticketNumber: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly role: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Guest: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly fullName: CodecTypes['pg/text@1']['input'];
      readonly phone: CodecTypes['pg/text@1']['input'] | null;
      readonly side: CodecTypes['pg/text@1']['input'];
      readonly rsvpStatus: CodecTypes['pg/text@1']['input'];
      readonly ticketNumber: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly role: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly guest: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly fullName: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly phone: CodecTypes['pg/text@1']['output'] | null;
      readonly rsvpStatus: CodecTypes['pg/text@1']['output'];
      readonly side: CodecTypes['pg/text@1']['output'];
      readonly ticketNumber: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly role: CodecTypes['pg/text@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly guest: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly fullName: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly phone: CodecTypes['pg/text@1']['input'] | null;
      readonly rsvpStatus: CodecTypes['pg/text@1']['input'];
      readonly side: CodecTypes['pg/text@1']['input'];
      readonly ticketNumber: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly role: CodecTypes['pg/text@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;